requirements.txt
Django==2.2.3
djangorestframework==вофтпщ3.9.4
gitpython==2.1.11